Create default gauge component
Connects to #34

Refactor DeviceStatusGauge to use default Gauge component

change-type: minor
